requirements.txt
click==7.1.2
Flask==1.1.2
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
sklearn==0.0
python-dotenv==0.16.0
requests==2.25.1
python-crontab==2.5.1
croniter==1.0.10
